# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SwiftShip Logistics Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed

  # Users
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read

  # Audit Logs
...
  /dashboard/order-trends:
    get:
      operationId: getOrderTrends
      tags: [dashboard]
      summary: Get order trends chart data
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Order trends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderTrendPoint"
...
  /dashboard/cod-summary:
    get:
      operationId: getCodSummary
      tags: [dashboard]
      summary: Get COD collection summary
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: COD summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodSummary"
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, manager, vendor, rider, staff]
        status:
          type: string
          enum: [active, inactive]
        stationId:
          type: ["integer", "null"]
        stationName:
          type: ["string", "null"]
        vendorId:
          type: ["integer", "null"]
        riderId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - status
        - createdAt
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
        relatedId:
          type: ["integer", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - title
        - message
        - type
        - isRead
        - createdAt
...
    OrderTrendPoint:
      type: object
      properties:
        date:
          type: string
        total:
          type: integer
        delivered:
          type: integer
        failed:
          type: integer
      required:
        - date
        - total
        - delivered
        - failed
...
    CodSummary:
      type: object
      properties:
        totalCodCollected:
          type: number
        totalDeliveryCharge:
          type: number
        totalVendorPayable:
          type: number
        totalReleased:
          type: number
        pendingRelease:
          type: number
      required:
        - totalCodCollected
        - totalDeliveryCharge
        - totalVendorPayable
        - totalReleased
        - pendingRelease