# lib/api-zod/src/generated/types/orderTrendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SwiftShip Logistics Management API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderTrendPoint {
  date: string;
  total: number;
  delivered: number;
  failed: number;
}